Split decimal header
